Split out PanelFocusManager
